File size can be long; don't truncate it to integer.
* cvs2svn_lib/cvs_revision.py (_parse_cvs_revision_state): Interpret
  file_size as a long, not an integer.


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@1999 be7e6eca-30d4-0310-a8e5-ac0d63af7087

+++ cvs2svn_lib/cvs_revision.py
@@ -265,7 +265,7 @@ def _parse_cvs_revision_state(line):
   file_in_attic = string_to_bool(file_in_attic)
   file_executable = string_to_bool(file_executable)
 
-  file_size = int(file_size)
+  file_size = long(file_size)
 
   deltatext_exists = string_to_bool(deltatext_exists)
